pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ner-sdk"
version = "0.1.0"
description = "Tiny SDK for rule-based NER labeling + JSON/JSONL IO"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["NER", "NLP", "dataset", "BIO", "JSONL"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://example.com/ner-sdk"

[project.scripts]
nersdk = "ner_sdk.cli:main"

[tool.setuptools.package-data]
"ner_sdk.domains.finance" = ["patterns.yaml"]
"ner_sdk.domains.medical" = ["patterns.yaml"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]